separated google logic in router into its own file
